Updated to allow deployment 'inplace' on the target system. Updated to use command line scp for updates

+++ build/SilverStripeDeployTask.php
@@ -1,6 +1,5 @@
 <?php
 
-
 /**
  * A phing task to perform a deployment of a specified
  * tarball to a remote server
@@ -9,66 +8,108 @@
  * @license BSD License http://silverstripe.org/bsd-license/
  */
 class SilverStripeDeployTask extends SilverStripeBuildTask {
-	
-	/* deployment config */ 
+	/* deployment config */
+
 	private $localpath;
-	
 	private $package = '';
-	
 	private $apachegroup = 'apache';
-	
 	private $remotepath = '';
-	
 	private $incremental = false;
-	
+
 	/* SSH configuration */
 	private $host = "";
-    private $port = 22;
-    private $username = "";
-    private $password = "";
-    private $pubkeyfile = '';
-    private $privkeyfile = '';
-    private $privkeyfilepassphrase = '';
+	private $port = 22;
+	private $username = "";
+	private $password = "";
+	private $pubkeyfile = '';
+	private $privkeyfile = '';
+	private $privkeyfilepassphrase = '';
 	private $ignoreerrors = false;
-	
+	// are we updating the /current release in-place, or creating a new release?
+	private $inplace = false;
+
 	public function main() {
-		
+
 		if (!strlen($this->pubkeyfile) && !strlen($this->password)) {
 			// prompt for the password
-			$this->password = $this->getInput("Password for ".$this->username.'@'.$this->host);
-		} 
+			$this->password = $this->getInput("Password for " . $this->username . '@' . $this->host);
+		}
 
 		$this->connect();
-		
-		$releasePath = $this->remotepath.'/releases/'.date('YmdHis');
-		$currentPath = $this->remotepath.'/current';
+
+		$currentPath = $this->remotepath . '/current';
+		$releasePath = $this->configureReleaseDir();
+
 		$remotePackage = $releasePath . '/' . $this->package;
 		$localPackage = $this->localpath . '/' . $this->package;
-		
-		$this->log("Configuring target directories at $releasePath");
-		$this->execute("mkdir --mode=2775 -p $releasePath/silverstripe-cache");
-		$this->execute("mkdir --mode=2775 -p $releasePath/assets");
 
 		$this->log("Copying deployment package $localPackage");
 		$this->copyFile($localPackage, $remotePackage);
-		
-		if ($this->incremental) {
+
+		$this->beforeDeploy($releasePath, $currentPath);
+		$this->extractPackage($remotePackage, $releasePath);
+		$this->doDeploy($releasePath, $currentPath);
+
+		@ssh2_exec($this->connection, 'exit');
+	}
+
+	/**
+	 * Configure the release directory in the remote system
+	 *
+	 * @return string 
+	 */
+	protected function configureReleaseDir() {
+		if ($this->inplace) {
+			$this->log("WARNING: Setting remote path equal to current deployment");
+			$releasePath = $this->remotepath . '/current';
+		} else {
+			$releasePath = $this->remotepath . '/releases/' . date('YmdHis');
+			$this->log("Configuring target directories at $releasePath");
+			$this->execute("mkdir --mode=2775 -p $releasePath/silverstripe-cache");
+			$this->execute("mkdir --mode=2775 -p $releasePath/assets");
+		}
+
+		return $releasePath;
+	}
+
+	/**
+	 * Copy the existing deployment and/or relevant configuration files
+	 * to the new location
+	 */
+	protected function beforeDeploy($releasePath, $currentPath) {
+		if ($this->incremental && !$this->inplace) {
 			$this->log("Copying existing deployment");
 			// we use rsync here to be able to use --excludes
 			$this->execute("rsync -r --exclude=silverstripe-cache $currentPath/* $releasePath/");
-			
+
 			$this->log("Copying configs");
 			$this->execute("cp $releasePath/mysite/.assets-htaccess $releasePath/assets/.htaccess");
 			$this->execute("cp $currentPath/.htaccess $releasePath/");
 			$this->execute("cp $currentPath/_ss_environment.php $releasePath/");
 			$this->execute("cp $currentPath/mysite/local.conf.php $releasePath/mysite/local.conf.php");
 		}
-		
+	}
+
+	/**
+	 * Extract the remote package
+	 *
+	 * @param type $remotePackage
+	 * @param type $releasePath 
+	 */
+	protected function extractPackage($remotePackage, $releasePath) {
 		$this->log("Extracting $remotePackage in $releasePath");
 		$this->execute("tar -zx -C $releasePath -f $remotePackage");
 		$this->execute("rm $remotePackage");
-		
-		if (!$this->incremental) {
+	}
+
+	/**
+	 * Update the new deployment with configs etc
+	 *
+	 * @param type $releasePath
+	 * @param type $currentPath 
+	 */
+	protected function doDeploy($releasePath, $currentPath) {
+		if (!$this->incremental && !$this->inplace) {
 			$this->log("Copying configs");
 			$this->execute("cp $releasePath/mysite/.assets-htaccess $releasePath/assets/.htaccess");
 			$this->execute("cp $currentPath/.htaccess $releasePath/");
@@ -78,139 +119,135 @@ class SilverStripeDeployTask extends SilverStripeBuildTask {
 			$this->log("Copying site assets");
 			$this->execute("cp -R $currentPath/assets $releasePath");
 		}
-		
+
 		$this->log("Backing up database");
 		$this->execute("php $currentPath/mysite/scripts/backup_database.php");
-		
+
 		$this->log("Executing dev/build");
 		$this->execute("php $releasePath/sapphire/cli-script.php dev/build");
 
-		$this->log("Changing symlinks");
-		$this->execute("rm $currentPath"); 
-		$this->execute("ln -s $releasePath $currentPath");
+		if (!$this->inplace) {
+			$this->log("Changing symlinks");
+			$this->execute("rm $currentPath");
+			$this->execute("ln -s $releasePath $currentPath");
+		}
 
 		$this->log("Fixing permissions");
 		$this->execute("chgrp -R $this->apachegroup $releasePath");
-		$this->execute("find $releasePath -type f -exec chmod 664 {} \;"); 
+		$this->execute("find $releasePath -type f -exec chmod 664 {} \;");
 		$this->execute("find $releasePath -type d -exec chmod 2775 {} \;");
 
 		$this->log("Finalising deployment");
 		$this->execute("touch $releasePath/DEPLOYED");
-		
-		@ssh2_exec($this->connection, 'exit');
 	}
-	
+
 	/**
 	 * Executes a command over SSH
 	 *
 	 * @param string $cmd
-	 *					The command to execute
-	 * @param boolean $canFail
-	 *					Whether it's okay for the command to fail when it's executed
 	 * @return string
 	 */
-	protected function execute($cmd, $canFail=false) {
-		$command = '('.$cmd.'  2>&1) && echo __COMPLETE';
+	protected function execute($cmd) {
+		$command = '(' . $cmd . '  2>&1) && echo __COMPLETE';
 		// $command = 'sh -c '.escapeshellarg('('.$this->command.'  2>&1) && echo __COMPLETE');
 
-        $stream = ssh2_exec($this->connection, $command);
-        if (!$stream) {
-            throw new BuildException("Could not execute command!");
-        }
+		$stream = ssh2_exec($this->connection, $command);
+		if (!$stream) {
+			throw new BuildException("Could not execute command!");
+		}
 
-        stream_set_blocking( $stream, true );
+		stream_set_blocking($stream, true);
 		$data = '';
-        while( $buf = fread($stream,4096) ){
-            $data .= $buf;
-        }
+		while ($buf = fread($stream, 4096)) {
+			$data .= $buf;
+		}
 
-		if (strpos($data, '__COMPLETE') !== false || $this->ignoreerrors || $canFail) {
+		if (strpos($data, '__COMPLETE') !== false || $this->ignoreerrors) {
 			$data = str_replace('__COMPLETE', '', $data);
 		} else {
 			$this->log("Command failed: $command", Project::MSG_WARN);
 			throw new BuildException("Failed executing command : $data");
 		}
-		
-        fclose($stream);
-		
+
+		fclose($stream);
+
 		return $data;
 	}
-	
+
 	/**
 	 * Copies a file to the remote system
 	 *
 	 * @param string $local
 	 * @param string $remote 
 	 */
-	protected function copyFile($localEndpoint, $remoteEndpoint)
-    {
+	protected function copyFile($localEndpoint, $remoteEndpoint) {
 		ssh2_sftp_mkdir($this->sftp, dirname($remoteEndpoint), 2775, true);
 		$ret = ssh2_scp_send($this->connection, $localEndpoint, $remoteEndpoint);
 
 		if ($ret === false) {
 			throw new BuildException("Could not create remote file '" . $remoteEndpoint . "'");
 		}
-    }
-	
+	}
+
 	/**
 	 * Connects SSH stuff up
 	 */
 	protected function connect() {
-		if (!function_exists('ssh2_connect')) { 
-            throw new BuildException("To use SshTask, you need to install the SSH extension.");
-        }
-		
-        $this->connection = ssh2_connect($this->host, $this->port);
-        if (is_null($this->connection)) {
-            throw new BuildException("Could not establish connection to " . $this->host . ":" . $this->port . "!");
-        }
+		if (!function_exists('ssh2_connect')) {
+			throw new BuildException("To use SshTask, you need to install the SSH extension.");
+		}
 
-        $could_auth = null;
-        if (strlen($this->pubkeyfile)) {
-            $could_auth = ssh2_auth_pubkey_file($this->connection, $this->username, $this->pubkeyfile, $this->privkeyfile, $this->privkeyfilepassphrase);
-        } else {
-            $could_auth = ssh2_auth_password($this->connection, $this->username, $this->password);
-        }
+		$this->connection = ssh2_connect($this->host, $this->port);
+		if (is_null($this->connection)) {
+			throw new BuildException("Could not establish connection to " . $this->host . ":" . $this->port . "!");
+		}
 
-        if (!$could_auth) {
-            throw new BuildException("Could not authenticate connection!");
-        }
-		
+		$could_auth = null;
+		if (strlen($this->pubkeyfile)) {
+			$could_auth = ssh2_auth_pubkey_file($this->connection, $this->username, $this->pubkeyfile, $this->privkeyfile, $this->privkeyfilepassphrase);
+		} else {
+			$could_auth = ssh2_auth_password($this->connection, $this->username, $this->password);
+		}
+
+		if (!$could_auth) {
+			throw new BuildException("Could not authenticate connection!");
+		}
+
 		$this->sftp = ssh2_sftp($this->connection);
 	}
-	
-	
+
 	public function setApachegroup($g) {
 		$this->apachegroup = $g;
 	}
+
 	public function getApachegroup() {
 		return $this->apachegroup;
 	}
-	
+
 	public function setLocalpath($p) {
 		$this->localpath = $p;
 	}
-	
+
 	public function getLocalpath() {
 		return $this->localpath;
 	}
-	
+
 	public function setRemotepath($p) {
 		$this->remotepath = $p;
 	}
-	
+
 	public function getRemotepath() {
 		return $this->remotepath;
 	}
-	
+
 	public function setPackage($pkg) {
 		$this->package = $pkg;
 	}
-	
+
 	public function getPackage() {
 		return $this->package;
 	}
-	
+
 	public function setIncremental($v) {
 		if (!is_bool($v)) {
 			$v = $v == 'true' || $v == 1;
@@ -218,116 +255,111 @@ class SilverStripeDeployTask extends SilverStripeBuildTask {
 
 		$this->incremental = $v;
 	}
-	
+
 	public function getIncremental() {
 		return $this->incremental;
 	}
-	
-	
-	public function setHost($host) 
-    {
-        $this->host = $host;
-    }
 
-    public function getHost() 
-    {
-        return $this->host;
-    }
+	public function setInplace($v) {
+		if (!is_bool($v)) {
+			$v = $v == 'true' || $v == 1;
+		}
 
-    public function setPort($port) 
-    {
+		$this->inplace = $v;
+	}
+
+	public function getInplace() {
+		return $this->inplace;
+	}
+
+	public function setHost($host) {
+		$this->host = $host;
+	}
+
+	public function getHost() {
+		return $this->host;
+	}
+
+	public function setPort($port) {
 		if (strpos($port, '${') === false) {
 			$this->port = $port;
 		}
-    }
+	}
 
-    public function getPort() 
-    {
-        return $this->port;
-    }
+	public function getPort() {
+		return $this->port;
+	}
 
-    public function setUsername($username) 
-    {
-        $this->username = $username;
-    }
+	public function setUsername($username) {
+		$this->username = $username;
+	}
 
-    public function getUsername() 
-    {
-        return $this->username;
-    }
+	public function getUsername() {
+		return $this->username;
+	}
 
-    public function setPassword($password) 
-    {
+	public function setPassword($password) {
 		if (strpos($password, '${') === false) {
 			$this->password = $password;
 		}
-    }
+	}
 
-    public function getPassword() 
-    {
-        return $this->password;
-    }
-	
-    /**
-     * Sets the public key file of the user to scp
-     */
-    public function setPubkeyfile($pubkeyfile)
-    {
+	public function getPassword() {
+		return $this->password;
+	}
+
+	/**
+	 * Sets the public key file of the user to scp
+	 */
+	public function setPubkeyfile($pubkeyfile) {
 		if (strpos($pubkeyfile, '${') === false) {
 			$this->pubkeyfile = $pubkeyfile;
 		}
-    }
+	}
 
-    /**
-     * Returns the pubkeyfile
-     */
-    public function getPubkeyfile()
-    {
-        return $this->pubkeyfile;
-    }
-    
-    /**
-     * Sets the private key file of the user to scp
-     */
-    public function setPrivkeyfile($privkeyfile)
-    {
-        $this->privkeyfile = $privkeyfile;
-    }
+	/**
+	 * Returns the pubkeyfile
+	 */
+	public function getPubkeyfile() {
+		return $this->pubkeyfile;
+	}
 
-    /**
-     * Returns the private keyfile
-     */
-    public function getPrivkeyfile()
-    {
-        return $this->privkeyfile;
-    }
-    
-    /**
-     * Sets the private key file passphrase of the user to scp
-     */
-    public function setPrivkeyfilepassphrase($privkeyfilepassphrase)
-    {
-        $this->privkeyfilepassphrase = $privkeyfilepassphrase;
-    }
+	/**
+	 * Sets the private key file of the user to scp
+	 */
+	public function setPrivkeyfile($privkeyfile) {
+		$this->privkeyfile = $privkeyfile;
+	}
 
-    /**
-     * Returns the private keyfile passphrase
-     */
-    public function getPrivkeyfilepassphrase($privkeyfilepassphrase)
-    {
-        return $this->privkeyfilepassphrase;
-    }
-    
-    public function setCommand($command) 
-    {
-        $this->command = $command;
-    }
+	/**
+	 * Returns the private keyfile
+	 */
+	public function getPrivkeyfile() {
+		return $this->privkeyfile;
+	}
 
-    public function getCommand() 
-    {
-        return $this->command;
-    }
-	
+	/**
+	 * Sets the private key file passphrase of the user to scp
+	 */
+	public function setPrivkeyfilepassphrase($privkeyfilepassphrase) {
+		$this->privkeyfilepassphrase = $privkeyfilepassphrase;
+	}
+
+	/**
+	 * Returns the private keyfile passphrase
+	 */
+	public function getPrivkeyfilepassphrase($privkeyfilepassphrase) {
+		return $this->privkeyfilepassphrase;
+	}
+
+	public function setCommand($command) {
+		$this->command = $command;
+	}
+
+	public function getCommand() {
+		return $this->command;
+	}
+
 	public function setIgnoreErrors($ignore) {
 		if (!is_bool($ignore)) {
 			$ignore = $ignore == 'true' || $ignore == 1;
@@ -335,8 +367,9 @@ class SilverStripeDeployTask extends SilverStripeBuildTask {
 
 		$this->ignoreerrors = $ignore;
 	}
-	
+
 	public function getIgnoreErrors() {
 		return $this->ignoreerrors;
 	}
+
 }
